Do not load design editor style rules inline
Design editor related styles are regular CSS, which do not need to be
loaded inline, as design editor does not use this and it does not use
liquid. Instead, load these styles to main.css, which is also minified.

As a result, delete the template-cs-style-rules file, as it is no longer
needed.